pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "unitechan"
version = "0.1.0"
description = "Pokemon Unite Team Splitting Bot"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.4.0,<3.0.0",
    "python-dotenv",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
